firmware/tcgbios: Prevent integer truncation in the return value
rc inside the function is uint32_t.  None of the callers check the return
value.

Coverity ID: 1055442

Acked-by: Ian Campbell <Ian.Campbell@citrix.com>
CC: Keir Fraser <keir@xen.org>
CC: Jan Beulich <JBeulich@suse.com>
CC: Ian Jackson <Ian.Jackson@eu.citrix.com>

+++ tools/firmware/rombios/32bit/tcgbios/tcgbios.c
@@ -428,7 +428,7 @@ uint32_t tcpa_get_laml(void)
  *  data       : pointer to the data (i.e., string) to be added to the log
  *  length     : length of the data
  */
-static uint16_t
+static uint32_t
 tcpa_add_measurement_to_log(uint32_t pcrIndex,
                             uint32_t event_type,
                             uint32_t event_id,
